Added Code Analogies to JavaScript
This is a website that takes a wide variety of complex topics in JS/CSS and shares simple/visual real-world analogies that represent them through articles and tutorials. This is my first ever pull request, as well as Hacktoberfest. Hopefully what I propose adds as much value to others as it has to me.