Update index.sass
Suggest adding a cursor pointer to the filter icon button in the GitHub issue description. This will enhance the user experience by providing a visual cue that the button is interactive and can be clicked. Users will be able to easily identify and use the filter functionality without the need for additional instructions or code changes.